Fix integration tests (#1239)
* Fix unit test code styles

* Fix integration tests

* Fix unit test

* Fix list jobs command

* Fix admin controller

* Rename admin controller to fl test driver

* Update readme.md

* Remove debug messages

+++ nvflare/app_common/homomorphic_encryption/he_intime_accumulate_model_aggregator.py
@@ -158,7 +158,7 @@ class HEInTimeAccumulateWeightedAggregator(Aggregator):
                 else:
                     self.warning_count[client_name] = 0
             n_iter = 1.0
-        float_n_iter = np.float(n_iter)
+        float_n_iter = np.float64(n_iter)
 
         aggregation_weight = self.aggregation_weights.get(client_name)
         if aggregation_weight is None:
